Use "int" instead of "long" for most things referring to 32-bit entities in cpuemu (for 64 bit compatibility), and changed the related printf()s.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1515 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/emu-i386/simx86/protmode.h
@@ -38,14 +38,14 @@
 #include "emu-ldt.h"
 
 typedef struct {
-	unsigned long BoundL,BoundH;
+	unsigned int BoundL,BoundH;
 	unsigned short Oldsel, Attrib __attribute__ ((packed));
 } SDTR;
 
 typedef struct {
-	unsigned long Base;
-	unsigned long Limit;
-	unsigned long Attrib;
+	unsigned int Base;
+	unsigned int Limit;
+	unsigned int Attrib;
 } DTR;
 
 /*
